+++ IntegrationTests/HealthCardControl/AuthenticateChallengeR2048Test.swift
@@ -42,7 +42,7 @@ final class AuthenticateChallengeR2048Test: CardSimulationTerminalTestCase {
         let challenge = "1234567890".data(using: .utf8)!
         _ = try await Self.healthCard
             .verify(pin: "123456", type: .mrpinHome)
-        let authenticatedResult = try await Self.healthCard.authenticate(challenge: challenge)
+        let authenticatedResult = try await Self.healthCard.authenticateAsync(challenge: challenge)
 
         expect(authenticatedResult.certificate.signatureAlgorithm) == .sha256RsaMgf1
         expect(authenticatedResult.certificate.certificate.count) == 1242