+++ Sources/HealthCardControl/SecureMessaging/CardType+SecureMessaging.swift
@@ -86,26 +86,26 @@ extension CardType {
     ///     - writeTimeout: time in seconds. Default: 30
     ///     - readTimeout: time in seconds. Default 30
     /// - Returns: SecureHealthCard
-    public func openSecureSession(
+    public func openSecureSessionAsync(
         can: CAN,
         writeTimeout: TimeInterval = 30,
         readTimeout: TimeInterval = 30
-    ) async throws -> HealthCardType {
+    ) async throws -> SecureHealthCardType {
         CommandLogger.commands.append(Command(message: "Open secure Session", type: .description))
         let channel = try openBasicChannel()
         // Read/Determine ApplicationIdentifier of the card's initial application
-        let cardAid = try await channel.determineCardAid(writeTimeout: writeTimeout, readTimeout: readTimeout)
+        let cardAid = try await channel.determineCardAidAsync(writeTimeout: writeTimeout, readTimeout: readTimeout)
 
         // Read EF.CardAccess and determine the algorithm for the key agreement (e.g. PACE)
-        let keyAgreementAlgorithm = try await channel.readKeyAgreementAlgorithm(cardAid: cardAid)
+        let keyAgreementAlgorithm = try await channel.readKeyAgreementAlgorithmAsync(cardAid: cardAid)
         // Read EF.Version2 and determine HealthCardPropertyType
-        let cardType = try await channel.readCardType(
+        let cardType = try await channel.readCardTypeAsync(
             cardAid: cardAid,
             writeTimeout: writeTimeout,
             readTimeout: readTimeout
         )
         let healthCard = try HealthCard(card: self, status: .valid(cardType: cardType))
-        let sessionKey = try await keyAgreementAlgorithm.negotiateSessionKey(
+        let sessionKey = try await keyAgreementAlgorithm.negotiateSessionKeyAsync(
             card: healthCard,
             can: can,
             writeTimeout: writeTimeout,
@@ -125,6 +125,7 @@ extension CardType {
     ///     - writeTimeout: time in seconds. Default: 30
     ///     - readTimeout: time in seconds. Default 30
     /// - Returns: Publisher that negotiates a secure session when scheduled to run.
+    @available(*, deprecated, message: "Use structured concurrency version instead")
     public func openSecureSession(can: String, writeTimeout: TimeInterval = 30, readTimeout: TimeInterval = 30)
         -> AnyPublisher<HealthCardType, Error> {
         let parsedCan: CAN
@@ -134,5 +135,29 @@ extension CardType {
             return Fail(error: error).eraseToAnyPublisher()
         }
         return openSecureSession(can: parsedCan, writeTimeout: writeTimeout, readTimeout: readTimeout)
+    }
+
+    /// Open a secure session with a Card for further scheduling/attaching Publisher commands
+    ///
+    /// - Note: The healthCard provided by the Combine operation chain should be used for the commands
+    ///   to be executed on the secure channel.
+    ///   After the chain has completed the session should be invalidated/closed.
+    ///
+    /// - Parameters:
+    ///     - can: The Channel access number for the session
+    ///     - writeTimeout: time in seconds. Default: 30
+    ///     - readTimeout: time in seconds. Default 30
+    /// - Returns: SecureHealthCard
+    public func openSecureSessionAsync(
+        can: String,
+        writeTimeout: TimeInterval = 30,
+        readTimeout: TimeInterval = 30
+    ) async throws -> SecureHealthCardType {
+        let parsedCan = try CAN.from(Data(can.utf8))
+        return try await openSecureSessionAsync(
+            can: parsedCan,
+            writeTimeout: writeTimeout,
+            readTimeout: readTimeout
+        )
     }
 }